[EUM-13] refactor: cursor util 공용화
