+++ src/modules/chat/services/room/room.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 import { AppException } from '../../../../common/errors/app.exception';
-import { decodeCursor, encodeCursor } from '../../utils/cursor.util';
+import { decodeRoomCursor, encodeCursor } from '../../utils/cursor.util';
 import { buildMessagePreview } from '../../utils/message-preview.util';
 
 import type {
@@ -163,10 +163,9 @@ export class RoomService {
     const me = BigInt(meUserId);
     const size = query.size ?? 20;
 
-    const cursor = query.cursor ? decodeCursor(query.cursor) : null;
+    const cursor = query.cursor ? decodeRoomCursor(query.cursor) : null;
     const cursorSortAt = cursor ? new Date(cursor.sortAt) : null;
-    const cursorRoomId =
-      cursor && 'roomId' in cursor ? BigInt(cursor.roomId) : null;
+    const cursorRoomId = cursor ? BigInt(cursor.roomId) : null;
 
     const myRoomIds = await this.participantRepo.getMyRoomIds(me);
     if (myRoomIds.length === 0) return { nextCursor: null, items: [] };
